Lambda/src/requirements.txt
boto3>=1.24.21
Pillow>=10.0.0
